+++ js/listener.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2020
+ * (c) Copyright Ascensio System SIA 2021
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -74,11 +74,6 @@
             return;
         }
         switch (event.data.method) {
-            case "editorShowHeaderButton":
-                if (OCA.Onlyoffice.ShowHeaderButton) {
-                    OCA.Onlyoffice.ShowHeaderButton();
-                }
-                break;
             case "editorRequestClose":
                 OCA.Onlyoffice.onRequestClose();
                 break;
@@ -104,5 +99,11 @@
                 break;
         }
     }, false);
+
+    window.addEventListener("popstate", function (event) {
+        if ($("#onlyofficeFrame").length) {
+            OCA.Onlyoffice.onRequestClose();
+        }
+    });
 
 })(OCA);